* hm-blender/master:
  Set filepath have a button now
  [UV Utility] fixes for the addon.
  -tab to spaces
  Bug Fix: render button.
  Code Cleanup
  Add Selective Remove Broken File.

+++ oscurart_tools/oscurart_render.py
@@ -3,22 +3,11 @@ import math
 import os
 
 
-## ------------- CHECK OVERRIDE LIST EXIST -----------------
-
-def checkOverridesExist():
-    for scene in bpy.data.scenes[:]:
-        try:
-            scene["OVERRIDE"]
-        except:
-            scene["OVERRIDE"] = "[]"
-
-
 ##-------------------------------- RENDER ALL SCENES ----------------------------
 
 
 def defRenderAll (frametype, scenes):
 
-    checkOverridesExist()
     
     activescene = bpy.context.scene
     FC = bpy.context.scene.frame_current
@@ -35,7 +24,7 @@ def defRenderAll (frametype, scenes):
     slotlist = { ob : [sl.material for sl in ob.material_slots] for ob in bpy.data.objects if ob.type in types if len (ob.material_slots)} 
 
     for scene in scenes:     
-        proptolist = list(eval(scene['OVERRIDE']))
+        proptolist = list(eval(scene.overrides))
         cursc = scene.name
         renpath = scene.render.filepath
         endpath = scene.render.filepath
@@ -129,7 +118,6 @@ bpy.types.Scene.rcPARTS = bpy.props.IntProperty(default=0, min=2, max=50, step=1
 
 def OscRenderCropFunc():
     
-    checkOverridesExist()
     
     SCENENAME = os.path.split(bpy.data.filepath)[-1].partition(".")[0]
     PARTS = bpy.context.scene.rcPARTS
@@ -155,7 +143,6 @@ class renderCrop (bpy.types.Operator):
 ##---------------------------BATCH MAKER------------------
 def defoscBatchMaker(TYPE):
     
-    checkOverridesExist() # overrides list exist
     
     if os.sys.platform.startswith("w"):
         print("PLATFORM: WINDOWS")
@@ -231,7 +218,6 @@ class oscBatchMaker (bpy.types.Operator):
 ## --------------------------------------PYTHON BATCH--------------------------------------------------------
 def defoscPythonBatchMaker(BATCHTYPE,SIZE):
     
-    checkOverridesExist() # overrides list exist
     
     # REVISO SISTEMA
     if os.sys.platform.startswith("w"):
@@ -314,6 +300,7 @@ class VarColArchivos (bpy.types.PropertyGroup):
     filename = bpy.props.StringProperty(name="", default="")
     value = bpy.props.IntProperty(name="", default=10)    
     fullpath = bpy.props.StringProperty(name="", default="")
+    checkbox = bpy.props.BoolProperty(name="", default=True)
 bpy.utils.register_class(VarColArchivos)
 
 class SumaFile(bpy.types.Operator):
@@ -330,7 +317,8 @@ class SumaFile(bpy.types.Operator):
                     i = bpy.context.scene.broken_files.add()
                     i.filename = f
                     i.fullpath = os.path.join(root,f)
-                    i.value = os.path.getsize(os.path.join(root,f))      
+                    i.value = os.path.getsize(os.path.join(root,f))   
+                    i.checkbox = True   
         return {'FINISHED'}  
 
 class ClearFile(bpy.types.Operator):
@@ -347,7 +335,8 @@ class DeleteFiles(bpy.types.Operator):
 
     def execute(self, context):
         for file in bpy.context.scene.broken_files:
-            os.remove(file.fullpath)
+            if file.checkbox:
+                os.remove(file.fullpath)
         bpy.context.scene.broken_files.clear()    
         return {'FINISHED'}     
 
@@ -370,6 +359,7 @@ class BrokenFramesPanel (bpy.types.Panel):
             colrow = col.row(align=1)
             colrow.prop(i, "filename")
             colrow.prop(i, "value")
+            colrow.prop(i, "checkbox")
 
         col = layout.column(align=1)
         colrow = col.row(align=1)
